chore(server): moving function
